+++ tests/snapshot_tests/test_snapshots.py
@@ -10,6 +10,7 @@ from tests.snapshot_tests.language_snippets import SNIPPETS
 from textual import events, on
 from textual.app import App, ComposeResult
 from textual.binding import Binding, Keymap
+from textual.command import SimpleCommand
 from textual.containers import Center, Container, Grid, Middle, Vertical, VerticalScroll
 from textual.pilot import Pilot
 from textual.renderables.gradient import LinearGradient
@@ -2751,6 +2752,23 @@ def test_select_width_auto(snap_compare):
     snap_compare(TallSelectApp(), run_before=run_before)
 
 
+
+def test_markup_command_list(snap_compare):
+    """Regression test for https://github.com/Textualize/textual/issues/5276
+    You should see a command list, with console markup applied to the action name and help text."""
+
+    class MyApp(App):
+        def on_mount(self) -> None:
+            self.search_commands(
+                [
+                    SimpleCommand(
+                        "Hello [u green]World", lambda: None, "Help [u red]text"
+                    )
+                ]
+            )
+
+    snap_compare(MyApp())
+
 def test_app_resize_order(snap_compare):
     """Regression test for https://github.com/Textualize/textual/issues/5284
     You should see a placeholder with text "BAR", focused and scrolled down so it fills the screen.
@@ -2791,3 +2809,4 @@ def test_app_resize_order(snap_compare):
             self.add_class("narrow")
 
     snap_compare(SCApp())
+